Add 1.4.py
This one was pretty interesting.
At first, it distracted me and maybe I got a little scared
when asking for a permutation of a palindrome.
But then I started thinking "What really is a palindrome?" and
I came to realize that a palindrome is a set of characters where
they can only come in pairs and only one letter is allowed to be
an odd number. You cannot have a palindrome with two odd number
sets.

For example
abccdcba -> Not a palindrome because neither side will ever be
even or reversible.